requirements.txt
# Core ML / Vision
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# HuggingFace (BioBERT / PubMedBERT / Tokenizers)
transformers>=4.37.0
tokenizers>=0.15.0
sentencepiece

# Data / Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0

# Image processing
pillow>=10.0.0

# Progress / utilities
tqdm>=4.66.0

# Explainability
shap>=0.44.0
captum>=0.7.0

# Serialisation
h5py

# Optional: faster tokenisation
sacremoses

# Web Application
streamlit>=1.32.0
plotly>=5.18.0
reportlab>=4.1.0
